myenv/DjangoApi/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==2.0.12
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.8
PyJWT==2.9.0
requests==2.26.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==1.26.20
